+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added `alert-primary` view
 - Added `destroy()` method to Popup view
 - Added `external` data option for `badge` views
+- Added `version` to library exports
+- Added Discovery.js credits into nav popup
 - Eliminated unnecessary query execution and view rendering on a report page when changing non-affecting page params. As a side effect page params `query`, `view`, `title`, `dzen` and `noedit` are not available for the main query and views anymore
 - Changes in `struct` view:
     - Added a view as table for array and object values